scripts/requirements.txt
supabase>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
sentence-transformers>=2.6.0,<3.0.0
tiktoken>=0.5.0,<1.0.0
numpy>=2.1.0